Update PackageManagerHelperTest test cases
