+++ lib/src/androidTest/java/com/whatw3words/androidwraper/PackageManagerHelperTest.kt
@@ -26,27 +26,6 @@ class PackageManagerHelperTest {
     }
 
     @Test
-    fun testSignatureDigest() {
-        // Read the byte array from the package_signature_bytes.txt file in the assets folder
-        val byteArray = context.assets.open("package_signature_bytes.txt")
-            .bufferedReader()
-            .readText()
-            .replace(" ", "")
-            .split(",")
-            .map { it.toByte() }
-            .toByteArray()
-
-        // Create a Signature object from the byte array
-        val signature = Signature(byteArray)
-
-        // Expected SHA1 signature
-        val expectedSHA1 = "79:57:BC:E7:48:D6:25:BC:8B:50:99:91:51:A3:8B:61:0F:F0:AF:23"
-
-        // Assert that the PackageManagerHelper.signatureDigest returns an SHA1 signature that matches the expected signature
-        assertEquals(expectedSHA1, PackageManagerHelper.signatureDigest(signature))
-    }
-
-    @Test
     fun testGetSignaturesCompatNotNull() {
         val packageManager = context.packageManager
         val packageName = context.packageName
