+++ lib/src/test/assets/package_signature_bytes.txt
